Sort of refactor codgen to not favor openapi
Most people using this codegen would be using it for CRDs, since the
default openapi is codegenned in advance and published. So let's not
treat CRDs as second-class usage.
