+++ generation/sources/from-cluster.sh
@@ -7,7 +7,7 @@ kubectl get --raw /openapi/v2 \
 deno run \
   --allow-read="generation/api-specs" \
   --allow-write="lib/builtin" \
-  generation/bin/emit-from-openapi.ts \
+  generation/cmd/emit-from-openapi.ts \
   "generation/api-specs/from-cluster.json" \
   "lib/builtin"
