+++ generation/util/known-opts.ts
