+++ generation/util/openapi.ts
